adding icons, and fixing up the grid and sizes in the project

+++ app/src/main/java/com/pizzapp/utilities/StaticFunctions.java
@@ -3,7 +3,7 @@ package com.pizzapp.utilities;
 import android.content.res.Resources;
 
 public class StaticFunctions {
-    public static int convertDpToPx(int dp)
+    public static int convertDpToPx(double dp)
     {
         return (int) (dp * Resources.getSystem().getDisplayMetrics().density);
     }
